Ju test and fix for getStatistic api

+++ src/main/java/att/mapper/MonthStatisticMapper.java
@@ -15,14 +15,14 @@ public interface MonthStatisticMapper {
 
     @Mapping(target = "tenantId", source = "context.tenantId")
     MonthlyUserStatisticInfoDto toDto(DataTimeContext<?> context, Integer userId, YearMonth yearMonth,
-                                      Integer workDays, Long overtimeHours, Long totalWorkHours,
+                                      Integer workDays, Double overtimeHours, Double totalWorkHours,
                                       Double vacationDays, Double sickDays);
 
     @Mapping(target = "monthStatisticKey.tenantId", source = "context.tenantId")
     @Mapping(target = "monthStatisticKey.idUser", source = "userId")
     @Mapping(target = "monthStatisticKey.monthStartDate", source = "monthStart")
     MonthStatistic toEntity(DataTimeContext<?> context, Integer userId, LocalDate monthStart,
-                            Integer workDays, Integer overtimeHours, Integer totalWorkHours,
+                            Integer workDays, Double overtimeHours, Double totalWorkHours,
                             Double vacationDays, Double sickDays);
 
 }
